rename `tray.rs` -> `systray.rs`
